requirements.txt
amqp==5.3.1
asgiref==3.9.1
billiard==4.2.1
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.10.4
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.0
django-cors-headers==4.7.0
djangorestframework==3.16.1
dotenv==0.9.9
idna==3.10
iniconfig==2.1.0
kombu==5.5.4
packaging==25.0
pluggy==1.6.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
Pygments==2.19.2
pytest==8.4.1
pytest-cov==6.2.1
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
redis==6.4.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
